Test and implement all function
Previously the all function was not actually being tested in
test_diary.py, the reason being that we were calling it as an attribute:

`diary.all` instead of calling it as a function with `diary.all()`.

Thgis reaulted in us getting the following error documented beneath:

When trying to test for equality between diary.all and a list of
instances of the DiaryEntry object
Getting the following error message:

```
  E       assert <bound method Diary.all of <lib.diary.Diary object at
  0x104028950>> == [<lib.diary_entry.DiaryEntry object at 0x10402 8990>,
  <lib.diary_entry.DiaryEntry object at 0x103fc7050>]
```

Not entriely sure how to make the quality in the tests.